+++ resources-ktx/CHANGELOG.md
@@ -1,5 +1,15 @@
 ## Unreleased
 
+## 1.3.0-0 (2021-06-27)
+
+### Dependencies
+
+- kotlin-stdlib-jdk8 1.4.32 -> kotlin-stdlib 1.5.20
+- androidx.appcompat 1.2.0 -> 1.3.0
+- androidx.annotation 1.1.0 -> 1.2.0
+- androidx.core 1.3.0 -> 1.5.0
+- androidx.fragment 1.3.0 -> 1.3.5
+
 ### Added
 
 - Wrapper `Text` to make it possible to work with plain `String` and `StringRes` in the same way.